Delegate createReadStream
